requirements.txt
fastapi
uvicorn
streamlit
requests
python-dotenv
pydantic
plotly
pandas
python-multipart